Changes by Balu

+++ src/main/java/me/idbi/hcf/Discord/LogLibrary.java
@@ -9,71 +9,85 @@ public class LogLibrary {
     private static final WebhookClientBuilder channel = SetupBot.builder;
 
     public static void sendKill(Player killer, Player victim) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("kill"))
             channel.buildJDA().send(Embeds.kill(killer, victim).build());
     }
 
     public static void sendDeath(Player died, Player killer, boolean fallDamage) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("death"))
             channel.buildJDA().send(Embeds.death(died, killer, fallDamage).build());
     }
 
     public static void sendFactionCreate(Player executor, String factionName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_create"))
             channel.buildJDA().send(Embeds.faction_create(executor, factionName).build());
     }
 
     public static void sendFactionDisband(Player executor, String factionName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_disband"))
             channel.buildJDA().send(Embeds.faction_disband(executor, factionName).build());
     }
 
     public static void sendJoin(Player joiner, Player inviter, String factionName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_join_leave"))
             channel.buildJDA().send(Embeds.faction_join(joiner, inviter, factionName).build());
     }
 
     public static void sendLeave(Player joiner, String factionName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_join_leave"))
             channel.buildJDA().send(Embeds.faction_leave(joiner, factionName).build());
     }
 
     public static void sendBank_withdraw(Player executor, int amount) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_bank"))
             channel.buildJDA().send(Embeds.faction_bank_withdraw(executor, amount).build());
     }
 
     public static void sendBank_deposit(Player executor, int amount) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_bank"))
             channel.buildJDA().send(Embeds.faction_bank_deposit(executor, amount).build());
     }
 
     public static void sendFaction_createRank(Player executor, String rankName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_rank"))
             channel.buildJDA().send(Embeds.faction_rank_create(executor, rankName).build());
     }
 
     public static void sendFaction_deleteRank(Player executor, String rankName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_rank"))
             channel.buildJDA().send(Embeds.faction_rank_delete(executor, rankName).build());
     }
 
     public static void sendRank_setpermission(Player executor, String rankName, String factionName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_rank"))
             channel.buildJDA().send(Embeds.faction_rank_setpermission(executor, rankName, factionName).build());
     }
 
     public static void sendRank_removepermission(Player executor, String rankName, String factionName) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("faction_rank"))
             channel.buildJDA().send(Embeds.faction_rank_removepermission(executor, rankName, factionName).build());
     }
 
     public static void sendAdminDuty(Player admin, boolean mode) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("admin_duty"))
             channel.buildJDA().send(Embeds.admin_duty(admin, mode).build());
     }
 
     public static void sendCommand(Player p, String command) {
+        if(channel == null) return;
         if (DiscordFile.getDiscord().getBoolean("commands"))
             channel.buildJDA().send(Embeds.commands(p, command).build());
     }
